+++ lib/src/clojure_lsp/shared.clj
@@ -549,6 +549,6 @@
       ;; normalize scheme:/some/path to scheme:///some/path
       (uri-encode (.getScheme uri) (.getPath uri)))))
 
-(defn ignore-path? [db path]
-  (let [paths-ignore-regex (get-in db [:settings :paths-ignore-regex] [])]
+(defn ignore-path? [settings path]
+  (let [paths-ignore-regex (set (get settings :paths-ignore-regex #{}))]
     (some #(re-matches (re-pattern %) (fs/unixify path)) paths-ignore-regex)))